ABRORAKA ZERO NI TOG'IRLENG VA PUL BERISHNI QILIB QOYING OYBEK AKA SIZ ESA TEST NI YOZVORING YANA

+++ Application/Services/AdminService.cs
@@ -28,31 +28,14 @@ public class AdminService(IUnitOfWork work,
         await _work.User.UpdateAsync(user);
     }
 
-    public async Task CreateCompanyAsync(string Company, string Creator)
-    {
-        var comp = new Company()
-        {
-            Name = Company,
-            CreatedAt = DateTime.Now,
-            Creator_Name = Creator,
-            Employees_Count = 0
-        };
-        await _work.Company.CreateAsync(comp);
-    }
-
-    public async Task DeleteCompanyAsync(int id)
-    {
-        var comp = await _work.Company.GetByIdAsync(id);
-        if (comp is null)
-            throw new StatusCodeExeption(HttpStatusCode.NotFound, "Comp is not found");
-        await _work.Company.DeleteAsync(comp);
-    }
-
     public async Task DeleteUserAsync(int id)
     {
         var user = await _work.User.GetByIdAsync(id);
         if (user is null)
             throw new StatusCodeExeption(HttpStatusCode.NotFound, "User not found!");
+
+        if (user.Roles == Role.SuperAdmin)
+            throw new StatusCodeExeption(HttpStatusCode.BadRequest, "XARRAKIRI");
 
         await _work.User.DeleteAsync(user);
     }   
@@ -61,9 +44,4 @@ public class AdminService(IUnitOfWork work,
         => (await _work.User.GetAllAsync())
             .Where(p =>  p.Roles == Role.Admin)
             .ToList();
-
-    public async Task<List<Company>> GetAllCompanies()
-    {
-        return await _work.Company.GetAllAsync();
-    }
 }
